style: serviceの初期化を集約
